requirements.txt
cython
spacy
bleach
html5lib
msgpack-numpy
tqdm
ujson
tensorboard
matplotlib
